Move helpers to accountUtils
